requirements.txt
asgiref==3.5.2
azure-core==1.26.0
azure-storage-blob==12.14.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.2
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
